// types/api-endpoints.d.ts
export declare type IdRequest = string | string
export declare type EmptyObject = Record<string, never>
export declare type UserObjectResponse =
  | {
      type: 'person'
      person: {
        email?: string
      }
      name: string | null
      avatar_url: string | null
      id: IdRequest
      object: 'user'
    }
  | {
      type: 'bot'
      bot:
        | EmptyObject
        | {
            owner:
              | {
                  type: 'user'
                  user:
                    | {
                        type: 'person'
                        person: {
                          email: string
                        }
                        name: string | null
                        avatar_url: string | null
                        id: IdRequest
                        object: 'user'
                      }
                    | {
                        id: IdRequest
                        object: 'user'
                      }
                }
              | {
                  type: 'workspace'
                  workspace: true
                }
          }
      name: string | null
      avatar_url: string | null
      id: IdRequest
      object: 'user'
    }
export declare type TextRequest = string
export declare type PartialUserObjectResponse =
  | {
      id: IdRequest
      object: 'user'
    }
  | UserObjectResponse
// ...
export declare type DateResponse = {
  start: string
  end: string | null
  time_zone: TimeZoneRequest | null
}
export declare type RichTextItemResponse =
  | {
      type: 'text'
      text: {
        content: string
        link: {
          url: TextRequest
        } | null
      }
      annotations: {
        bold: boolean
        italic: boolean
        strikethrough: boolean
        underline: boolean
        code: boolean
        color:
          | 'default'
          | 'gray'
          | 'brown'
          | 'orange'
          | 'yellow'
          | 'green'
          | 'blue'
          | 'purple'
          | 'pink'
          | 'red'
          | 'gray_background'
          | 'brown_background'
          | 'orange_background'
          | 'yellow_background'
          | 'green_background'
          | 'blue_background'
          | 'purple_background'
          | 'pink_background'
          | 'red_background'
      }
      plain_text: string
      href: string | null
    }
  | {
      type: 'mention'
      mention:
        | {
            type: 'user'
            user: PartialUserObjectResponse
          }
        | {
            type: 'date'
            date: DateResponse
          }
        | {
            type: 'link_preview'
            link_preview: {
              url: TextRequest
            }
          }
        | {
            type: 'page'
            page: {
              id: IdRequest
            }
          }
        | {
            type: 'database'
            database: {
              id: IdRequest
            }
          }
      annotations: {
        bold: boolean
        italic: boolean
        strikethrough: boolean
        underline: boolean
        code: boolean
        color:
          | 'default'
          | 'gray'
          | 'brown'
          | 'orange'
          | 'yellow'
          | 'green'
          | 'blue'
          | 'purple'
          | 'pink'
          | 'red'
          | 'gray_background'
          | 'brown_background'
          | 'orange_background'
          | 'yellow_background'
          | 'green_background'
          | 'blue_background'
          | 'purple_background'
          | 'pink_background'
          | 'red_background'
      }
      plain_text: string
      href: string | null
    }
  | {
      type: 'equation'
      equation: {
        expression: TextRequest
      }
      annotations: {
        bold: boolean
        italic: boolean
        strikethrough: boolean
        underline: boolean
        code: boolean
        color:
          | 'default'
          | 'gray'
          | 'brown'
          | 'orange'
          | 'yellow'
          | 'green'
          | 'blue'
          | 'purple'
          | 'pink'
          | 'red'
          | 'gray_background'
          | 'brown_background'
          | 'orange_background'
          | 'yellow_background'
          | 'green_background'
          | 'blue_background'
          | 'purple_background'
          | 'pink_background'
          | 'red_background'
      }
      plain_text: string
      href: string | null
    }